# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Hotel:
    properties:
      address:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /example/helloworld:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - example
  /hotels:
    post:
      consumes:
      - application/json
      description: Add a hotel to the database
      parameters:
      - description: Hotel
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/models.Hotel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: Add a hotel
  /hotels/{name}:
    get:
      description: Get a hotel from the database by its name
      parameters:
      - description: Hotel Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: Get a hotel by name
    put:
      consumes:
      - application/json
      description: Update a hotel in the database
      parameters:
      - description: Hotel Name
        in: path
        name: name
        required: true
        type: string
      - description: Hotel
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/models.Hotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Hotel'
      summary: Update a hotel
swagger: "2.0"